add JS glue for Wallet management
