+++ js/Tx.js
@@ -170,41 +170,6 @@ export const verify = (module, config, tx, xpub, signature) => {
         return result === 0
 };
 
-export const createWallet = (module, seed) => {
-    const input_str = JSON.stringify(seed);
-    const input_array = iconv.encode(input_str, 'utf8');
-
-    const bufinput  = newArray(module, input_array);
-    const bufoutput = newArray0(module, 4096);
-
-    let rsz = module.xwallet_create(bufinput, input_array.length, bufoutput);
-    let output_array = copyArray(module, bufoutput, rsz);
-
-    module.dealloc(bufoutput);
-    module.dealloc(bufinput);
-
-    let output_str = iconv.decode(Buffer.from(output_array), 'utf8');
-    return JSON.parse(output_str);
-};
-
-export const spend = (module, wallet, inputs, outputs, fee_addr) => {
-    const input = { wallet: wallet, inputs: inputs, outputs: outputs, fee_addr: fee_addr};
-    const input_str = JSON.stringify(input);
-    const input_array = iconv.encode(input_str, 'utf8');
-
-    const bufinput  = newArray(module, input_array);
-    const bufoutput = newArray0(module, 4096);
-
-    let rsz = module.xwallet_spend(bufinput, input_array.length, bufoutput);
-    let output_array = copyArray(module, bufoutput, rsz);
-
-    module.dealloc(bufoutput);
-    module.dealloc(bufinput);
-
-    let output_str = iconv.decode(Buffer.from(output_array), 'utf8');
-    return JSON.parse(output_str);
-};
-
 export default {
   newTxOut: apply(newTxOut, RustModule),
   newTxIn:  apply(newTxIn, RustModule),
@@ -213,6 +178,4 @@ export default {
   addOutput: apply(addOutput, RustModule),
   sign: apply(sign, RustModule),
   verify: apply(verify, RustModule),
-  createWallet: apply(createWallet, RustModule),
-  spend: apply(spend, RustModule)
 };
